requirements.txt
agora-token-builder==1.0.0
asgiref==3.6.0
Django==4.1.4
gunicorn==20.1.0
python-dotenv==0.21.0
sqlparse==0.4.3
